Split integration tests by command
